Renamed package from `app` to `core`.
