+++ app/src/main/java/com/example/myapplication/core/presentation/theme/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.app
+package com.example.myapplication.core.presentation.theme.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.myapplication.app.ui.theme.MyApplicationTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
