+++ app/src/main/java/com/example/myapplication/core/presentation/theme/ui/Type.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.app.ui.theme
+package com.example.myapplication.core.presentation.theme.ui
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
